krs #24 - Moved DAOs to more appropriate packages
- Packaged by implemenetation, not by type
