+++ src/main/java/org/kingsski/api/config/dao/CachingDaoConfig.java
@@ -2,11 +2,11 @@ package org.kingsski.api.config.dao;
 
 import org.kingsski.api.config.DaoConfig;
 import org.kingsski.api.config.DaoDelegateConfig;
-import org.kingsski.api.dao.individual.IndividualDao;
-import org.kingsski.api.dao.race.CachingRaceDao;
-import org.kingsski.api.dao.race.RaceDao;
-import org.kingsski.api.dao.team.CachingTeamDao;
-import org.kingsski.api.dao.team.TeamDao;
+import org.kingsski.api.dao.IndividualDao;
+import org.kingsski.api.dao.caching.CachingRaceDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.caching.CachingTeamDao;
+import org.kingsski.api.dao.TeamDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
