+++ src/main/java/org/kingsski/api/config/dao/NonCachingDaoConfig.java
@@ -2,9 +2,9 @@ package org.kingsski.api.config.dao;
 
 import org.kingsski.api.config.DaoConfig;
 import org.kingsski.api.config.DaoDelegateConfig;
-import org.kingsski.api.dao.individual.IndividualDao;
-import org.kingsski.api.dao.race.RaceDao;
-import org.kingsski.api.dao.team.TeamDao;
+import org.kingsski.api.dao.IndividualDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.TeamDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
