+++ src/main/java/org/kingsski/api/config/dao/delegate/DummyDaoDelegateConfig.java
@@ -1,14 +1,12 @@
 package org.kingsski.api.config.dao.delegate;
 
 import org.kingsski.api.config.DaoDelegateConfig;
-import org.kingsski.api.dao.individual.IndividualDao;
-import org.kingsski.api.dao.individual.StubIndividualDao;
-import org.kingsski.api.dao.race.DummyRaceDao;
-import org.kingsski.api.dao.race.RaceDao;
-import org.kingsski.api.dao.race.StubRaceDao;
-import org.kingsski.api.dao.team.DummyTeamDao;
-import org.kingsski.api.dao.team.StubTeamDao;
-import org.kingsski.api.dao.team.TeamDao;
+import org.kingsski.api.dao.IndividualDao;
+import org.kingsski.api.dao.stub.StubIndividualDao;
+import org.kingsski.api.dao.dummy.DummyRaceDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.dummy.DummyTeamDao;
+import org.kingsski.api.dao.TeamDao;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
