+++ src/main/java/org/kingsski/api/config/dao/delegate/RepositoryDaoDelegateConfig.java
@@ -1,13 +1,13 @@
 package org.kingsski.api.config.dao.delegate;
 
 import org.kingsski.api.config.DaoDelegateConfig;
-import org.kingsski.api.dao.individual.IndividualDao;
-import org.kingsski.api.dao.individual.StubIndividualDao;
-import org.kingsski.api.dao.race.RaceDao;
-import org.kingsski.api.dao.race.StubRaceDao;
-import org.kingsski.api.dao.team.RepositoryTeamDao;
-import org.kingsski.api.dao.team.TeamDao;
-import org.kingsski.api.dao.team.TeamRepository;
+import org.kingsski.api.dao.IndividualDao;
+import org.kingsski.api.dao.stub.StubIndividualDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.stub.StubRaceDao;
+import org.kingsski.api.dao.repository.RepositoryTeamDao;
+import org.kingsski.api.dao.TeamDao;
+import org.kingsski.api.dao.repository.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
