+++ src/main/java/org/kingsski/api/config/dao/delegate/StubDaoDelegateConfig.java
@@ -1,12 +1,12 @@
 package org.kingsski.api.config.dao.delegate;
 
 import org.kingsski.api.config.DaoDelegateConfig;
-import org.kingsski.api.dao.individual.IndividualDao;
-import org.kingsski.api.dao.individual.StubIndividualDao;
-import org.kingsski.api.dao.race.RaceDao;
-import org.kingsski.api.dao.race.StubRaceDao;
-import org.kingsski.api.dao.team.StubTeamDao;
-import org.kingsski.api.dao.team.TeamDao;
+import org.kingsski.api.dao.IndividualDao;
+import org.kingsski.api.dao.stub.StubIndividualDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.stub.StubRaceDao;
+import org.kingsski.api.dao.stub.StubTeamDao;
+import org.kingsski.api.dao.TeamDao;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
