+++ src/main/java/org/kingsski/api/dao/dummy/DummyTeamDao.java
@@ -1,5 +1,6 @@
-package org.kingsski.api.dao.team;
+package org.kingsski.api.dao.dummy;
 
+import org.kingsski.api.dao.TeamDao;
 import org.kingsski.api.model.Team;
 
 import java.util.ArrayList;
