+++ src/test/java/org/kingsski/api/dao/race/CachingRaceDaoTest.java
@@ -3,8 +3,8 @@ package org.kingsski.api.dao.race;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kingsski.api.dao.race.CachingRaceDao;
-import org.kingsski.api.dao.race.RaceDao;
+import org.kingsski.api.dao.RaceDao;
+import org.kingsski.api.dao.caching.CachingRaceDao;
 import org.kingsski.api.model.Race;
 import org.mockito.Spy;
 import org.mockito.runners.MockitoJUnitRunner;
