+++ src/test/java/org/kingsski/api/dao/team/CachingTeamDaoTest.java
@@ -3,8 +3,8 @@ package org.kingsski.api.dao.team;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kingsski.api.dao.team.CachingTeamDao;
-import org.kingsski.api.dao.team.TeamDao;
+import org.kingsski.api.dao.TeamDao;
+import org.kingsski.api.dao.caching.CachingTeamDao;
 import org.kingsski.api.model.Team;
 import org.mockito.Spy;
 import org.mockito.runners.MockitoJUnitRunner;
