+++ src/test/java/org/kingsski/api/dao/team/RepositoryTeamDaoTest.java
@@ -3,6 +3,8 @@ package org.kingsski.api.dao.team;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.kingsski.api.dao.repository.RepositoryTeamDao;
+import org.kingsski.api.dao.repository.TeamRepository;
 import org.kingsski.api.model.Team;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
